requirements.txt
numpy==1.23
opencv-python
Pillow
tensorboard
gin-config
dm_pix
rawpy
mediapy
flatdict
pytorch-lightning==1.5.2
opencv-python==4.5.4.58
matplotlib==3.5.0
imageio==2.10.4
scipy==1.8.0
imageio-ffmpeg==0.4.5
tqdm==4.64.0